# swagger_server/models/trip_update.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class TripUpdate(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, start_date: date=None, end_date: date=None, ship: int=None, season: int=None):  # noqa: E501
        """TripUpdate - a model defined in Swagger

        :param id: The id of this TripUpdate.  # noqa: E501
        :type id: int
        :param start_date: The start_date of this TripUpdate.  # noqa: E501
        :type start_date: date
        :param end_date: The end_date of this TripUpdate.  # noqa: E501
        :type end_date: date
        :param ship: The ship of this TripUpdate.  # noqa: E501
        :type ship: int
        :param season: The season of this TripUpdate.  # noqa: E501
        :type season: int
        """
        self.swagger_types = {
            'id': int,
            'start_date': date,
            'end_date': date,
            'ship': int,
            'season': int
        }

        self.attribute_map = {
            'id': 'id',
            'start_date': 'startDate',
            'end_date': 'endDate',
            'ship': 'ship',
            'season': 'season'
        }
        self._id = id
        self._start_date = start_date
        self._end_date = end_date
        self._ship = ship
        self._season = season

    @classmethod
    def from_dict(cls, dikt) -> 'TripUpdate':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The trip_update of this TripUpdate.  # noqa: E501
        :rtype: TripUpdate
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this TripUpdate.


        :return: The id of this TripUpdate.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this TripUpdate.


        :param id: The id of this TripUpdate.
        :type id: int
        """

        self._id = id

    @property
    def start_date(self) -> date:
        """Gets the start_date of this TripUpdate.


        :return: The start_date of this TripUpdate.
        :rtype: date
        """
        return self._start_date

    @start_date.setter
    def start_date(self, start_date: date):
        """Sets the start_date of this TripUpdate.


        :param start_date: The start_date of this TripUpdate.
        :type start_date: date
        """
        if start_date is None:
            raise ValueError("Invalid value for `start_date`, must not be `None`")  # noqa: E501

        self._start_date = start_date

    @property
    def end_date(self) -> date:
        """Gets the end_date of this TripUpdate.


        :return: The end_date of this TripUpdate.
        :rtype: date
        """
        return self._end_date

    @end_date.setter
    def end_date(self, end_date: date):
        """Sets the end_date of this TripUpdate.


        :param end_date: The end_date of this TripUpdate.
        :type end_date: date
        """
        if end_date is None:
            raise ValueError("Invalid value for `end_date`, must not be `None`")  # noqa: E501

        self._end_date = end_date

    @property
    def ship(self) -> int:
        """Gets the ship of this TripUpdate.


        :return: The ship of this TripUpdate.
        :rtype: int
        """
        return self._ship

    @ship.setter
    def ship(self, ship: int):
        """Sets the ship of this TripUpdate.


        :param ship: The ship of this TripUpdate.
        :type ship: int
        """

        self._ship = ship

    @property
    def season(self) -> int:
        """Gets the season of this TripUpdate.


        :return: The season of this TripUpdate.
        :rtype: int
        """
        return self._season

    @season.setter
    def season(self, season: int):
        """Sets the season of this TripUpdate.


        :param season: The season of this TripUpdate.
        :type season: int
        """

        self._season = season